Fix Bug in Test Assertion
Need to train myself better on red-green-refactor. The previous assert
always passed, because I wasn't triggering an actual Mock method. I had
to change to verifying that the call is in the list of calls. I've
verified that it fails when the call isn't made from the code under
test and that it passes when it is.