Factor Out Supporting Fixtures Into conftest.py
Yeah, I ended up going back to this. I still don't enjoy having this
code so far separate from the code, but I'm hoping as long as it is
in the same directory, it won't be too hard to discover. All of the
extra code was making it hard to see the tests for the trees.