Authenticate.perform() Returns Success Integer
In order to pipe Authenticate into a CLI program, I wanted a way for
outside systems to verify that everything went well. In order to do
that I had to return an integer from the authentication process.
This required me to move code logic out of __init__ (yeah, the only
reason it was there was because it was the simplest way at the time)
and put it into a separate perform() method.