Introduce System Test For Post-Auth User Command
When this feature is implemented, the program can be passed a
LETS_DO_POSTCMD environment variable that identifies a
program/script that will be called during the post-authentication
stage of the script. This command will be passed the environment
variables from the lets-do-dns program, which can be used to trigger
automated reloading of configuration on a daemon relying upon the
newly generated SSL certificate.