Tweak coverage.py operations
Have coverage.py show all inspected files, even on success. Also,
have coverage.py not report a branch problem in Authenticate(). The
reason why he was reporting that the elif branch couldn't get hit
was because of self._in_cleanup_hook_stage calling an attribute on
the instance's passed environment parameter. I believe coverage.py,
not knowing whether the attribute was legit, made a faulty
assumption that the branch would never be called. Changing this to
an if cleans up this log, as both if checks (currently), cannot be
true at the same time.