Major fixture modifications
Remove almost every fixture from the unit tests and replace with
simplified MagicMock()s or plain text/integer assignments in the
tests. The code feels much easier to read now in these tests. While
the fixtures seemed like a good idea at the time, they added
complexity and logic to the tests, and were overly stringent in
what behaviour had to be present for the tests to pass.