Cleanup variable names and stub_int stub
The stub_int variable was introduced and fed into a autospec'd
patch() to make the test_create_stores_record_id_internally test
method to prepare for some changes that are about to happen in the
Record() class.

I also changed some variable names from having a stub prefix to
having a mock prefix.