Move record creation verification call to Authenticate()
While it makes sense to keep the main verification logic in
Record(), the more I looked at the code, the more convinced I became
that having extensive logic surrounding triggering and repeating of
this verification in Record() doesn't "feel" right.

The idea is to keep the core functionality of verifying whether a
Record() instance exists in DNS in Record(), but triggering this and
looping through it should exist in Authenticate(), at least for now.