Factor out API operation error handling
Both create() and delete() contained the exact same logic, with the
only difference being which requests operation was being called, and
one of the keyword arguments passed to that operation. After much
contemplation, I determined a method that seems to me to be readable
that allowed me to factor this logic out into a private method.