When command.run() fails, include output text in exception
When command.run('program string') fails, we will include the
STDOUT and STDERR streamed from the program in the exception error
message. This is very helpful when attempting to debug why a
program passed in the LETS_DO_POSTCMD environment variable fails,
as you can send error output to STDOUT/STDERR and be able to inspect
it later. certbot will even output this text, since lets-do-dns will
fail with a non-zero exit code in this scenario.