Merge Record() functionality into Authenticate()
Record() seemed to be functioning as a wrapper for other
objects/functions in the system, and it's purpose seemed redundant to
the point that I couldn't remember if something required Record() or
Resource(). Due to this, I decided it was time to pull the small amount
of unique functionality into Authenticate() and rely on directly
calling Resource() instead of proxying messages to it.